third_party/WebKit/LayoutTests/http/tests/credentialmanager/passwordcredential-basics.html

<!DOCTYPE html>
<title>Credential Manager: PasswordCredential basics.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="/serviceworker/resources/interfaces.js"></script>
<script>
test(function() {
    var credential = new PasswordCredential({
        id: 'id',
        password: 'pencil',
        name: 'name',
        iconURL: 'https://example.com/icon.png'
    });

    verify_interface('PasswordCredential', credential, {
        id: 'string',
        name: 'string',
        iconURL: 'string',
        type: 'string'
    });

    assert_equals(credential.id, 'id');
    assert_equals(credential.name, 'name');
    assert_equals(credential.iconURL, 'https://example.com/icon.png');
    assert_equals(credential.type, 'password');
    assert_equals(credential.idName, 'username');
    assert_equals(credential.passwordName, 'password');
    assert_equals(credential.additionalData, null);
}, 'Interfaces and attributes of PasswordCredential');

test(function() {
    assert_throws(new SyntaxError(), function () {
        var credential = new PasswordCredential({
            id: 'id',
            password: 'pencil',
            name: 'name',
            iconURL: '-'
        });
    });
}, 'Construct a PasswordCredential with an invalid icon URL.');

test(function() {
    var credential = new PasswordCredential({
        id: 'id',
        password: 'pencil',
        name: 'name',
    });

    assert_equals(credential.id, 'id');
    assert_equals(credential.name, 'name');
    assert_equals(credential.iconURL, '');
    assert_equals(credential.type, 'password');
    assert_equals(credential.idName, 'username');
    assert_equals(credential.passwordName, 'password');
    assert_equals(credential.additionalData, null);

}, 'Construct a PasswordCredential with an empty icon URL.');

test(function() {
    var credential = new PasswordCredential({
        id: 'id',
        password: 'pencil',
    });

    assert_equals(credential.id, 'id');
    assert_equals(credential.name, '');
    assert_equals(credential.iconURL, '');
    assert_equals(credential.type, 'password');
    assert_equals(credential.idName, 'username');
    assert_equals(credential.passwordName, 'password');
    assert_equals(credential.additionalData, null);
}, 'Construct a PasswordCredential with an empty name and icon URL.');

test(function() {
    var credential = new PasswordCredential({
        id: 'id',
        password: 'pencil',
    });

    credential.idName = 'yay';
    assert_equals(credential.idName, 'yay');

    credential.passwordName = 'boo';
    assert_equals(credential.passwordName, 'boo');

    var additionalData = new FormData();
    credential.additionalData = additionalData;
    assert_equals(credential.additionalData, additionalData);
}, 'Verify the basics of "idName", "passwordName", and "additionalData"');
</script>
